backend/requirements.txt
channels==4.0.0
daphne==4.1.2
Django==5.1
django-channels==0.7.0
aiohttp==3.11.8
requests==2.31.0
urllib3==1.26.16
asgiref==3.8.1
httpx == 0.28.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1